% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{count_mutation_types}
\alias{count_mutation_types}
\title{Count mutation types}
\usage{
count_mutation_types(mut_types)
}
\arguments{
\item{mut_types}{A vector of MutationType objects}
}
\value{
A dataframe containing the count of each mutation
type
}
\description{
Given a list of objects of class MutationType, it counts
the occurrences of each one of them by grouping together
all mutation types having the same UP, DOWN, REF and ALT
sequences.
It returns a data frame containing two columns, "MutationType"
and "Count".
}
\examples{
m1 <- MutationType(up=Biostrings::DNAString("AA"), down=Biostrings::DNAString("CC"), ref=Biostrings::DNAString("C"), alt=Biostrings::DNAString("G"))
m2 <- MutationType(up=Biostrings::DNAString("TA"), down=Biostrings::DNAString("GA"), ref=Biostrings::DNAString("T"), alt=Biostrings::DNAString("C"))
mut_types_list <- list(m1, m2)
count_mutation_types(mut_types_list)
}
