% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_types.R
\name{get_mutation_types}
\alias{get_mutation_types}
\title{Get mutation types}
\usage{
get_mutation_types(vcf_file, ref_genome, context_length)
}
\arguments{
\item{vcf_file}{A .vcf or .vcf.gz file, loaded in memory}

\item{ref_genome}{A reference genome. hg19, hg38 and mm10
are supported}

\item{context_length}{The number of nucleotides to retrieve
for the upstream and downstream sequences on ref_genome}
}
\value{
A vector of MutationTypes
}
\description{
Given a VCF file, a string representing
a reference genome, and an integer for the context_length,
it returns a vector of objects of the MutationType class.
Each MutationType will contain information about the mutated
nucleotide as well as its upstream and downstream sequences
(up to <context_length> nucleotides of distance from the mutated base).
Mutation types are obtained by allowing only C/T nucleotides
as the reference ones. If that was not the case in the VCF file,
the complementary nucleotide is considered. The same holds for the
upstream/downstream sequences taken from the reference genome, that
are reverse complemented if needed.
}
\details{
Compressed VCF files are supported.
reference_genome has to be a string containing an
official genome name.
}
\examples{
vcf_file <- system.file("extdata", "ex2.vcf", package="VariantAnnotation")
get_mutation_types(vcf_file, 'hg19', 3)
}
