% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\docType{class}
\name{Mutation-class}
\alias{Mutation-class}
\alias{Mutation}
\title{An S4 class to represent a Mutation.}
\description{
An S4 class to represent a Mutation.
}
\section{Slots}{

\describe{
\item{\code{chromosome}}{The chromosome where the mutation is located}

\item{\code{position}}{The position of the mutation on the chromosome}

\item{\code{ref}}{The reference base}

\item{\code{alt}}{The alternative base}
}}

\examples{
Mutation(chromosome="21", position=123456789, ref=Biostrings::DNAString("C"), alt=Biostrings::DNAString("G"))
}
