% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\docType{class}
\name{MutationType-class}
\alias{MutationType-class}
\alias{MutationType}
\title{An S4 class to represent a mutation type.}
\description{
An S4 class to represent a mutation type.
}
\section{Slots}{

\describe{
\item{\code{up}}{The upstream sequence, from the reference genome}

\item{\code{down}}{The downstream sequence, from the reference genome}

\item{\code{ref}}{The reference base}

\item{\code{alt}}{The alternative base                                                                                        ref="DNAString", alt="DNAString"))}
}}

\examples{
MutationType(up=Biostrings::DNAString("AA"), down=Biostrings::DNAString("CC"), ref=Biostrings::DNAString("C"), alt=Biostrings::DNAString("G"))
}
